# docs/backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Renda+ ERP — API do servidor
  version: v1
  description: |
    Contrato da API usada pelo app do Mac (Sprints 1 a 4).
    Regras gerais:
    - Toda rota exige sessão (`Authorization: Bearer <token>`), exceto `GET /api/v1/status` e `POST /api/v1/session` (ADR-005).
      Sem sessão válida: 401 `UNAUTHENTICATED`. Perfil sem a permissão: 403 `FORBIDDEN` (a negativa fica na auditoria).
    - Edição exige `If-Match` com a versão lida (ETag); versão desatualizada: 412 `VERSION_MISMATCH`; sem cabeçalho: 428.
    - Cadastro exige `Idempotency-Key` (8 a 100 caracteres); repetir com a mesma chave devolve o mesmo registro;
      mesma chave com outro conteúdo: 422 `IDEMPOTENCY_KEY_REUSED`.
    - Cliente e fornecedor são papéis do mesmo parceiro, cada um com a sua situação. Cadastrar num papel o CNPJ de um parceiro
      que ainda não o tem devolve 422 `PARTNER_OTHER_ROLE`, com `details` `partnerId` e `version` para chamar `.../{id}/enable`.
    - Valores decimais (custo de referência, fator de conversão) trafegam como texto com ponto, até 6 casas (ADR-006, PD-007).
    - Valores em reais trafegam em centavos, como texto de inteiro (`"30207146"` = R$ 302.071,46); datas de negócio `AAAA-MM-DD`.
    - Documentos comerciais (Sprint 4): a revisão emitida da proposta é imutável; o pedido só muda em rascunho. Confirmar o pedido
      (`Idempotency-Key` + `If-Match`) cria o projeto, um equipamento por unidade das linhas de equipamento e um título a receber
...
servers:
...
security:
  - sessao: []
...
paths:
...
  /api/v1/units-of-measure:
    get:
      tags: [Unidades e categorias]
      summary: Unidades de medida, ativas e inativas (item.read)
      responses:
        '200': { description: Unidades, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/UnitOfMeasure' } } } } }
    post:
      tags: [Unidades e categorias]
      summary: Inclui unidade (catalog.admin); código de 1 a 10 letras ou algarismos, guardado em maiúsculas
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [code, name], properties: { code: { type: string, maxLength: 10 }, name: { type: string, maxLength: 60 } } } } }
      responses:
        '201': { description: Incluída, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/UnitOfMeasure' } } } }
        '403': { $ref: '#/components/responses/Erro' }
        '422': { $ref: '#/components/responses/Erro' }
  /api/v1/units-of-measure/{code}:
    parameters: [{ name: code, in: path, required: true, schema: { type: string } }]
    put:
      tags: [Unidades e categorias]
      summary: Renomeia ou inativa (catalog.admin); o código não muda
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [name, status], properties: { name: { type: string, maxLength: 60 }, status: { type: string, enum: [ATIVO, INATIVO] } } } } }
      responses:
        '200': { description: Nova versão, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/UnitOfMeasure' } } } }
        '403': { $ref: '#/components/responses/Erro' }
        '412': { $ref: '#/components/responses/Erro' }
        '422': { $ref: '#/components/responses/Erro' }
  /api/v1/item-categories:
    get:
      tags: [Unidades e categorias]
      summary: Categorias de item, ativas e inativas (item.read)
      responses:
        '200': { description: Categorias, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ItemCategory' } } } } }
    post:
      tags: [Unidades e categorias]
      summary: Inclui categoria (catalog.admin); nome único sem diferenciar maiúsculas
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string, maxLength: 100 } } } } }
      responses:
        '201': { description: Incluída, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ItemCategory' } } } }
        '403': { $ref: '#/components/responses/Erro' }
        '422': { $ref: '#/components/responses/Erro' }
  /api/v1/item-categories/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      tags: [Unidades e categorias]
      summary: Renomeia ou inativa (catalog.admin)
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [name, status], properties: { name: { type: string, maxLength: 100 }, status: { type: string, enum: [ATIVO, INATIVO] } } } } }
      responses:
        '200': { description: Nova versão, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ItemCategory' } } } }
        '403': { $ref: '#/components/responses/Erro' }
        '412': { $ref: '#/components/responses/Erro' }
        '422': { $ref: '#/components/responses/Erro' }
  /api/v1/proposals:
    get:
      tags: [Propostas]
      summary: Lista de propostas (proposal.read), com a revisão vigente
      parameters:
        - { name: search, in: query, required: false, schema: { type: string }, description: 'Número, título ou cliente' }
        - { name: status, in: query, required: false, schema: { type: string, enum: [ABERTA, GANHA, PERDIDA, TODOS] } }
      responses:
        '200': { description: Propostas, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ProposalSummary' } } } } }
    post:
      tags: [Propostas]
      summary: Nova proposta com a revisão 1 em rascunho (proposal.create), idempotente; número PR00001
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProposalRequest' } } }
      responses:
        '201': { description: Cadastrada, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Proposal' } } } }
        '403': { $ref: '#/components/responses/Erro' }
        '422': { $ref: '#/components/responses/Erro' }
  /api/v1/proposals/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Propostas]
      summary: Proposta com todas as revisões e linhas (proposal.read)
      responses:
        '200': { description: Proposta, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Proposal' } } } }
        '404': { $ref: '#/components/responses/Erro' }
    put:
      tags: [Propostas]
      summary: Altera a revisão em rascunho (proposal.update); revisão emitida devolve 409
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProposalRequest' } } }
      responses:
        '200': { description: Nova versão, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Proposal' } } } }
        '409': { $ref: '#/components/responses/Erro' }
        '412': { $ref: '#/components/responses/Erro' }
        '422': { $ref: '#/components/responses/Erro' }
        '428': { $ref: '#/components/responses/Erro' }
  /api/v1/proposals/{id}/issue:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Propostas]
      summary: Emite a revisão em rascunho — IssueProposalRevision (proposal.issue); emitida fica imutável
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      responses:
        '200': { description: Revisão emitida, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Proposal' } } } }
        '409': { $ref: '#/components/responses/Erro' }
        '412': { $ref: '#/components/responses/Erro' }
  /api/v1/proposals/{id}/revisions:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Propostas]
      summary: Cria a próxima revisão em rascunho, copiada da última emitida (proposal.update); as anteriores ficam preservadas
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      responses:
        '200': { description: Nova revisão, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Proposal' } } } }
        '409': { $ref: '#/components/responses/Erro' }
        '412': { $ref: '#/components/responses/Erro' }
  /api/v1/proposals/{id}/outcome:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Propostas]
      summary: Registra a perda com motivo — RecordProposalOutcome (proposal.update); o ganho é a conversão em pedido
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/OutcomeRequest' } } }
      responses:
        '200': { description: Proposta perdida, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Proposal' } } } }
        '409': { $ref: '#/components/responses/Erro' }
        '412': { $ref: '#/components/responses/Erro' }
        '422': { $ref: '#/components/responses/Erro' }
  /api/v1/proposals/{id}/orders:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Propostas]
      summary: Converte a revisão emitida em pedido em rascunho — ConvertProposalToOrder (sales_order.create), idempotente; a proposta fica ganha
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: false
        content: { application/json: { schema: { $ref: '#/components/schemas/ConvertRequest' } } }
      responses:
        '201': { description: Pedido em rascunho, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/SalesOrder' } } } }
        '409': { $ref: '#/components/responses/Erro' }
        '422': { $ref: '#/components/responses/Erro' }
  /api/v1/proposals/{id}/history:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Propostas]
      summary: Histórico da proposta (proposal.read)
      responses:
        '200': { description: Histórico, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/HistoryEntry' } } } } }
...
components:
  securitySchemes:
    sessao:
      type: http
      scheme: bearer
      description: Token opaco devolvido por POST /api/v1/session
  headers:
    ETag:
      description: Versão do registro entre aspas, para usar no If-Match
      schema: { type: string, example: '"3"' }
  parameters:
    Id:
      { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    IfMatch:
      { name: If-Match, in: header, required: true, schema: { type: string, example: '"3"' }, description: 'Versão lida (ETag)' }
    IdempotencyKey:
      { name: Idempotency-Key, in: header, required: true, schema: { type: string, pattern: '^[A-Za-z0-9_-]{8,100}$' } }
  responses:
    Erro:
      description: Erro padronizado
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
  schemas:
    ApiError:
      type: object
      properties:
        code: { type: string, example: VERSION_MISMATCH }
        message: { type: string }
        details: { type: array, items: { type: object, properties: { field: { type: [string, 'null'] }, message: { type: string } } } }
        correlationId: { type: string }
        retryable: { type: boolean }
...
    HistoryEntry:
      type: object
      properties:
        occurredAt: { type: string, format: date-time }
        actor: { type: string }
        action:
          type: string
          description: >-
            PARTNER_REGISTERED, PARTNER_UPDATED, PARTNER_DEACTIVATED, PARTNER_ROLE_ENABLED, ITEM_REGISTERED, ITEM_UPDATED,
            ITEM_DEACTIVATED, PROPOSAL_DRAFTED, PROPOSAL_UPDATED, PROPOSAL_REVISION_ISSUED, PROPOSAL_REVISION_CREATED,
            PROPOSAL_LOST, PROPOSAL_WON, SALES_ORDER_DRAFTED, SALES_ORDER_UPDATED, SALES_ORDER_CONFIRMED, SALES_ORDER_CANCELLED,
            PROJECT_CREATED, PROJECT_CLOSED, EQUIPMENT_CREATED, EQUIPMENT_UPDATED, EQUIPMENT_CANCELLED
        version: { type: integer }
        reason: { type: [string, 'null'] }
        changes:
          type: object
          additionalProperties:
            type: object
            properties:
              before: { type: [string, 'null'] }
              after: { type: [string, 'null'] }
...
    UnitOfMeasure:
      type: object
      properties:
        code: { type: string, example: KG }
        name: { type: string }
        status: { type: string, enum: [ATIVO, INATIVO] }
        version: { type: string }
        items: { type: integer, description: 'Itens que usam a unidade' }
        updatedAt: { type: [string, 'null'], format: date-time }
        updatedBy: { type: [string, 'null'] }
    ItemCategory:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { type: string, enum: [ATIVO, INATIVO] }
        version: { type: string }
        items: { type: integer }
        suppliers: { type: integer, description: 'Fornecedores que a fornecem' }
        updatedAt: { type: [string, 'null'], format: date-time }
        updatedBy: { type: [string, 'null'] }
    Cents:
      type: string
      pattern: '^\d{1,18}$'
      description: Valor em centavos, texto de inteiro ("30207146" = R$ 302.071,46)
    SalesLineRequest:
      type: object
      required: [kind, quantity, unitPrice]
      properties:
        id: { type: [string, 'null'], description: 'Id da linha existente; vazio numa linha nova' }
        kind: { type: string, enum: [EQUIPAMENTO, MATERIAL, SERVICO] }
        itemId: { type: [string, 'null'], format: uuid, description: 'Obrigatório para MATERIAL e SERVICO (mesma natureza); vazio para EQUIPAMENTO' }
        description: { type: [string, 'null'], maxLength: 200, description: 'Modelo do equipamento; nos itens, a descrição do cadastro quando vazio' }
        quantity: { type: string, example: '10.5', description: 'Até 6 casas; equipamento em unidades inteiras (1 a 100)' }
        unitPrice: { type: string, example: '16.33', description: 'Preço unitário em reais, até 6 casas' }
        discountCents: { type: [string, 'null'], example: '10000' }
    SalesLine:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [EQUIPAMENTO, MATERIAL, SERVICO] }
        itemId: { type: [string, 'null'], format: uuid }
        itemCode: { type: [string, 'null'] }
        description: { type: string }
        quantity: { type: string }
        uom: { type: string }
        unitPrice: { type: string }
        discountCents: { $ref: '#/components/schemas/Cents' }
        grossCents: { $ref: '#/components/schemas/Cents' }
        totalCents: { $ref: '#/components/schemas/Cents', description: 'round(quantidade × preço) − desconto, meio-par por linha (PD-002)' }
    ProposalRequest:
      type: object
      required: [customerId, title, validUntil, lines]
      properties:
        customerId: { type: string, format: uuid }
        unitId: { type: [string, 'null'], format: uuid }
        title: { type: string, maxLength: 200 }
        validUntil: { type: string, format: date }
        paymentTerms: { type: [string, 'null'], maxLength: 500 }
        lines: { type: array, items: { $ref: '#/components/schemas/SalesLineRequest' } }
    ProposalRevision:
      type: object
      properties:
        id: { type: string, format: uuid }
        revision: { type: integer }
        status: { type: string, enum: [RASCUNHO, EMITIDA] }
        validUntil: { type: string, format: date }
        paymentTerms: { type: [string, 'null'] }
        totalCents: { $ref: '#/components/schemas/Cents' }
        issuedAt: { type: [string, 'null'], format: date-time }
        issuedBy: { type: [string, 'null'] }
        lines: { type: array, items: { $ref: '#/components/schemas/SalesLine' } }
    Proposal:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: PR00001 }
        customerId: { type: string, format: uuid }
        customerCode: { type: string }
        customerName: { type: string }
        unitId: { type: [string, 'null'], format: uuid }
        unitName: { type: [string, 'null'] }
        title: { type: string }
        status: { type: string, enum: [ABERTA, GANHA, PERDIDA] }
        outcomeReason: { type: [string, 'null'] }
        currentRevision: { type: integer }
        revisions: { type: array, items: { $ref: '#/components/schemas/ProposalRevision' } }
        version: { type: string }
        createdAt: { type: string, format: date-time }
        createdBy: { type: string }
        updatedAt: { type: [string, 'null'], format: date-time }
        updatedBy: { type: [string, 'null'] }
    ProposalSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        customerCode: { type: string }
        customerName: { type: string }
        unitName: { type: [string, 'null'] }
        title: { type: string }
        status: { type: string, enum: [ABERTA, GANHA, PERDIDA] }
        revision: { type: integer }
        revisionStatus: { type: string, enum: [RASCUNHO, EMITIDA] }
        validUntil: { type: string, format: date }
        totalCents: { $ref: '#/components/schemas/Cents' }
        version: { type: string }
    OutcomeRequest:
      type: object
      required: [outcome, reason]
      properties:
        outcome: { type: string, enum: [PERDIDA] }
        reason: { type: string, maxLength: 500 }
    ConvertRequest:
      type: object
      properties:
        unitId: { type: [string, 'null'], format: uuid, description: 'Obrigatória se a proposta não tem unidade' }
        contractDate: { type: [string, 'null'], format: date, description: 'Hoje quando vazia' }
...
    Installment:
      type: object
      required: [dueDate, amountCents]
      properties:
        seq: { type: [integer, 'null'], description: 'Na resposta: sequência a partir de 1 (a ordem da lista)' }
        dueDate: { type: string, format: date }
        amountCents: { $ref: '#/components/schemas/Cents' }
        milestone: { type: [string, 'null'], maxLength: 200 }
...
    SalesOrder:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: PV00001 }
        customerId: { type: string, format: uuid }
        customerCode: { type: string }
        customerName: { type: string }
        unitId: { type: string, format: uuid }
        unitName: { type: string }
        proposalId: { type: [string, 'null'], format: uuid }
        proposalCode: { type: [string, 'null'] }
        proposalRevision: { type: [integer, 'null'] }
        contractDate: { type: string, format: date }
        promisedDate: { type: [string, 'null'], format: date }
        notes: { type: [string, 'null'] }
        status: { type: string, enum: [DRAFT, CONFIRMED, IN_EXECUTION, COMPLETED, CANCELLED] }
        totalCents: { $ref: '#/components/schemas/Cents' }
        scheduledCents: { $ref: '#/components/schemas/Cents' }
        lines: { type: array, items: { $ref: '#/components/schemas/SalesLine' } }
        installments: { type: array, items: { $ref: '#/components/schemas/Installment' } }
        confirmedAt: { type: [string, 'null'], format: date-time }
        confirmedBy: { type: [string, 'null'] }
        projectId: { type: [string, 'null'], format: uuid }
        projectCode: { type: [string, 'null'] }
        projectStage: { type: [string, 'null'] }
        equipment:
          type: array
          items: { type: object, properties: { id: { type: string }, code: { type: string }, model: { type: string }, serialNumber: { type: [string, 'null'] }, status: { type: string } } }
        titles:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              code: { type: string }
              label: { type: string }
              dueDate: { type: string, format: date }
              competence: { type: string, example: '2026-10' }
              originalCents: { $ref: '#/components/schemas/Cents' }
              balanceCents: { $ref: '#/components/schemas/Cents' }
              status: { type: string, enum: [OPEN, PARTIAL, SETTLED, RENEGOTIATED, CANCELLED] }
        cancelledAt: { type: [string, 'null'], format: date-time }
        cancelledBy: { type: [string, 'null'] }
        cancelReason: { type: [string, 'null'] }
        version: { type: string }
        createdAt: { type: string, format: date-time }
        createdBy: { type: string }
        updatedAt: { type: [string, 'null'], format: date-time }
        updatedBy: { type: [string, 'null'] }
...